daemon: extract host-auth closeout from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the
host-authorization closeout owners, their bounded runner, the outcome
summarizer, and the on-cancel closeout move verbatim into
daemon_apply_hostauth.go.

Moved declarations (byte-identical bodies, no rename, no logic change):
hostAuthOwner, hostAuthOwnerOutcome, hostAuthCloseoutBudget,
hostAuthCloseoutOwners, runHostAuthCloseoutOwners,
summarizeHostAuthCloseout, applyHostAuthorizationCloseout,
closeoutHostAuthOnCancel. Same package, no caller changes; per-file
imports pruned by goimports. Build: go build ./pkg/daemon/... green.
